remove redundant ResponseBody interface and move ErrorMessage to DTO package
