+++ src/main/java/ru/yandex/practicum/filmorate/dto/ErrorMessage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.model;
+package ru.yandex.practicum.filmorate.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public final class ErrorMessage implements ResponseBody {
+public final class ErrorMessage {
 
     private String message;
