--- src/main/java/ru/yandex/practicum/filmorate/model/ResponseBody.java
@@ -1,5 +1,0 @@
-package ru.yandex.practicum.filmorate.model;
-
-public sealed interface ResponseBody permits User, Film, ErrorMessage {
-
-}
